updates that create the post successfully

+++ internal/database/model/airingSchedule.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type AiringSchedule struct {
-	ID          int          `json:"-" gorm:"column:id;primaryKey"`                            // Primary key for the airing schedule
+	ID          uint         `json:"-" gorm:"column:id;primaryKey"`                            // Primary key for the airing schedule
 	ExternalId  int          `json:"id" gorm:"column:external_id;unique"`                      // Unique external ID corresponding to Anilist
-	AiringAt    int          `json:"airingAt" gorm:"column:airing_at"`                         // Airing time
+	AiringAt    int64        `json:"airingAt" gorm:"column:airing_at"`                         // Airing time
 	Episode     int          `json:"episode" gorm:"column:episode"`                            // Episode number
 	MediaId     uint         `json:"mediaId" gorm:"column:media_id"`                           // Foreign key to the media table
 	Media       Media        `json:"media,omitempty" gorm:"foreignKey:MediaId"`                // Related media data
